reorganize the project structure as requested by the mentor - add new assignedUserId variable at Task.kt
