+++ src/main/kotlin/data/repositories/AuthenticationRepositoryImpl.kt
@@ -1,0 +1,6 @@
+package org.qudus.squad.data.repositories
+
+import org.qudus.squad.logic.repositories.AuthenticationRepository
+
+class AuthenticationRepositoryImpl : AuthenticationRepository {
+}
