+++ src/main/kotlin/data/repositories/TaskRepositoryImpl.kt
@@ -1,0 +1,6 @@
+package org.qudus.squad.data.repositories
+
+import org.qudus.squad.logic.repositories.TaskRepository
+
+class TaskRepositoryImpl: TaskRepository {
+}
