--- src/main/kotlin/data/repositories/authentication_repository/AuthenticationRepositoryImpl.kt
@@ -1,6 +1,0 @@
-package org.qudus.squad.data.repositories.authentication_repository
-
-import org.qudus.squad.logic.repositories.authentication_repository.AuthenticationRepository
-
-class AuthenticationRepositoryImpl : AuthenticationRepository {
-}
