--- src/main/kotlin/data/repositories/project_repository/ProjectRepositoryImpl.kt
@@ -1,6 +1,0 @@
-package org.qudus.squad.data.repositories.project_repository
-
-import org.qudus.squad.logic.repositories.project_repository.ProjectRepository
-
-class ProjectRepositoryImpl: ProjectRepository {
-}
