--- src/main/kotlin/data/repositories/task_repository/TaskRepositoryImpl.kt
@@ -1,6 +1,0 @@
-package org.qudus.squad.data.repositories.task_repository
-
-import org.qudus.squad.logic.repositories.task_repository.TaskRepository
-
-class TaskRepositoryImpl: TaskRepository {
-}
