+++ src/main/kotlin/logic/repositories/TaskRepository.kt
@@ -1,0 +1,4 @@
+package org.qudus.squad.logic.repositories
+
+interface TaskRepository {
+}
