--- src/main/kotlin/logic/repositories/authentication_repository/AuthenticationRepository.kt
@@ -1,4 +1,0 @@
-package org.qudus.squad.logic.repositories.authentication_repository
-
-interface AuthenticationRepository {
-}
