--- src/main/kotlin/logic/repositories/project_repository/ProjectRepository.kt
@@ -1,4 +1,0 @@
-package org.qudus.squad.logic.repositories.project_repository
-
-interface ProjectRepository {
-}
