--- src/main/kotlin/logic/repositories/task_repository/TaskRepository.kt
@@ -1,4 +1,0 @@
-package org.qudus.squad.logic.repositories.task_repository
-
-interface TaskRepository {
-}
