+++ src/main/kotlin/logic/usecases/CreateNewProjectUseCase.kt
@@ -1,0 +1,8 @@
+package org.qudus.squad.logic.usecases
+
+import org.qudus.squad.logic.repositories.ProjectRepository
+
+class CreateNewProjectUseCase(
+    private val projectRepository: ProjectRepository
+) {
+}
