--- src/main/kotlin/logic/usecases/authentication/emptyFile.kt
@@ -1,2 +1,0 @@
-package org.qudus.squad.logic.usecases.authentication
-
