--- src/main/kotlin/logic/usecases/log/emptyFile.kt
@@ -1,2 +1,0 @@
-package org.qudus.squad.logic.usecases.log
-
