--- src/main/kotlin/logic/usecases/project/CreateNewProjectUseCase.kt
@@ -1,8 +1,0 @@
-package org.qudus.squad.logic.usecases.project
-
-import org.qudus.squad.logic.repositories.project_repository.ProjectRepository
-
-class CreateNewProjectUseCase(
-    private val projectRepository: ProjectRepository
-) {
-}
