--- src/main/kotlin/logic/usecases/task/emptyFile.kt
@@ -1,2 +1,0 @@
-package org.qudus.squad.logic.usecases.task
-
