+++ src/main/kotlin/ui/features/CreateNewProjectUI.kt
@@ -1,0 +1,8 @@
+package org.qudus.squad.ui.features
+
+import org.qudus.squad.logic.usecases.CreateNewProjectUseCase
+
+class CreateNewProjectUI(
+    private val createNewProjectUseCase: CreateNewProjectUseCase
+) {
+}
