--- src/main/kotlin/ui/features/log/emptyFile.kt
@@ -1,2 +1,0 @@
-package org.qudus.squad.ui.features.log
-
