--- src/main/kotlin/ui/features/project/CreateNewProjectUI.kt
@@ -1,8 +1,0 @@
-package org.qudus.squad.ui.features.project
-
-import org.qudus.squad.logic.usecases.project.CreateNewProjectUseCase
-
-class CreateNewProjectUI(
-    private val createNewProjectUseCase: CreateNewProjectUseCase
-) {
-}
